rename package names, rename application class name as well
