+++ src/test/java/com/encrypter/enc/EncApplicationTests.java
@@ -1,10 +1,10 @@
-package com.encrypter.enc.backend;
+package com.encrypter.enc;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EncBackendApplicationTests {
+class EncApplicationTests {
 
     @Test
     void contextLoads() {
